Update base image to Debian Trixie
This commit updates the base image used in the development container from
the specific SHA256 hash of the Debian Bookworm image to the latest Debian
Trixie image. This change ensures that the development environment is up-
to-date with the latest stable release, which may include important
security updates, bug fixes, and new features.

Signed-off-by: John Strunk <john.strunk@gmail.com>