Add corepack enable and install yarn command
This commit adds the command to enable corepack and install yarn globally
as a post-create command in the devcontainer.json file. This ensures that
corepack is enabled and yarn is installed when the development container is
created, streamlining the setup process for new developers. Additionally,
the command to install the Gemini CLI globally has been added to the post-
create commands to facilitate its use in the development environment.

Signed-off-by: John Strunk <john.strunk@gmail.com>